requirements.txt
requests
python-nmap
python-dotenv
